add two fizzbuzz examples
one example uses the "if" function
and the other uses a "select" function
provided by the example.

It demonstrates why select isn't really possible
in this version of the language due to the purely lexical nature of scoping.

Experimentation showed that merging in the calling environment
works as a hackish workaround.

There were also minor changes to some edge cases in the language.